Replace usage of system `time` command with `resource` module.
The use of the `time` command w/ `subprocess.Popen` _and_ w/out `shell=True` will cause Python to try to execute a process named `time` on the system. Many (most?) systems do not have a system `time` command present, instead relying on the shell built-in command. This PR removes the use of the `time` command and instead uses the Python `resource` module to measure system and user CPU time for children processes of the `run_benchmarks.py` process. In this case, the only child process is the one started by `subprocess.Popen`, so we are able to get the CPU time for that command specifically, just as if `time` were used.

Related: openjournals/joss-reviews#4991