Make pyperf tests faster (#19)
The pyperf tests are very slow. However if the benchmark is run on a properly isolated system, this many repetitions is overkill. We therefore use the `--fast` flag to reduce the number of tests.

The size of the FD-Laplace test is also increased to be closer to real-world problems and more similar to the size of the other problems in the benchmark. The convergence criterion for that test is also fixed, and the function now returns the number of iterations.

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>