Reduce code duplication and add Numba for Python 3.11 (#21)
As of yesterday Numba now supports Python 3.11. This means that Python 3.11 no longer needs to be treated as a special case. Therefore we add Numba tests to Python 3.11. Code duplication in the workflows is vastly decreases by using a matrix strategy.

Additionally fixes #22.

The latest version of Pythran (0.13.1) is now used in the benchmarks.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>