Use new Pyccel command as of Version 2 (#40)
The `pyccel` command has received substantial modifications after the
release of Version 2. In particular, it has a new flag
`--compiler-family` instead of `--compiler-laguage`, and the flag
`--export-compile-info` was renamed as `--export-compiler-config`.

We make here the necessary modifications to:
- the GitHub action `generate_pyccel_config`
- the GitHub action `create_readme`
- the Python script `benchmarks/run_benchmarks.py`
- the README instructions

Further, we update the filenames in the GitHub workflow `run_benchmarks`
(Python 3.11 is used rather than 3.10).

---------

Co-authored-by: Emily Bourne <louise.bourne@gmail.com>