Update README.md (#43)
In PR #40 the file `README.md` was updated by hand, forgetting that it
is generated by the action `create_readme` using the files `intro.md`
and `test_details.md`. As a consequence, the changes made to `README.md`
were reverted by the first workflow run. Here we finally
update`intro.md`, and move it to a folder called `readme` together with
`test_details.md`.